Change the response format from logger output to a structured JSON response
By using a JSON response format, this allows uniqush clients to know if
the API was successful.
API clients will be able to easily determine the reason an API failed,
without resorting to heuristics (checking if the response contains a
certain string).
Additionally, it is much easier for clients to process the outcome of a
push, when a push involves more than one device.

-   bump major version from 1.5.x to 2.0.0, since the API response format
    changed, and some users may be parsing the response.

(This is mostly the same as what we use internally.
 The original author(David Jeong)'s commits were edited to remove a field that
wasn't used by upstream.
 The commits were rebased and squashed after PR #87 was approved)

Barely related to issue #55 (Related in the sense that the responses
are much easier to parse)