Commit /subscriptions, /services and /rebuildserviceset API endpoints
/subscriptions shows the subscription details (tokens/regids, push
service type, service names) of a given subscriber. If no service is
provided in the query, all services are queried.

As part of this change, the set of all service names must be tracked in
redis. KEYS would be too slow to use, so a new set SERVICES.{0} was added.

- If the redis databases sharded, then the {0} could be used as a shard scheme,
  to avoid hitting multiple databases for what would be a small set.
- /rebuildserviceset is added, to be manually run to create the service
  set for pre-existing uniqush users. This hasn't been fully tested yet.

For convenience, /psps was also added.
- The API response maps service names to a list of psps with that service name.
  (From the documentation, I'm not sure if it is meant to be possible
   for a service name to be shareable among 2 or more PSPs, e.g. with
   different push service types.)

Add a config for the new logging types; fix logging bugs