Migrate the redis implementation to redis.v3, Remove invalid PSPs
redis.v3 is better tested and has sentinel support.
This implementation has been in use at our company for a while.

I'd expect it to have better performance, since it has better error
handling and doesn't shut down connections unnecessarily after an mget.

Automatically remove invalid PSPs if they are seen
(A reference to a key which is successfully requested, but redis
tells us it does not exist)