Handle unsubscribes with the uniqush binary API
The initial version failed to send unsubscribe updates on resChan,
which is something the binary API would do.
The binary provider API sends status codes to push_service.go
(From a byte of the TCP response with the associated message id)

In order for uniqush to properly unsubscribe, it must send the same
information over resChan, so that push_service.go can unsubscribe.

Long term, the abstraction would probably be flipped
(Make an HTTPClient interface out of the binary provider API)

Add constant definitions for the binary API's bytes.