Fix some more notices noticed by gometalinter
NOTE: The old regex for services and subscribers was wrong.
Fix it so that `a-z, A-Z, 0-9, -, _, @ or .` are actually the only
accepted characters.

Temporarily explicitly allow `\`, `^`, `[`, and `]` so that any old
clients will continue to work.
Forbid backticks because I doubt they'd be used.
(uniqush.org didn't mention valid characters for APIs)

TODO: Emit deprecation warnings in the server logs and in API responses in another PR.